# Optimize WeakCache