--- a/Orm/Xtensive.Orm/Orm/Internals/FieldAccessors/EntityFieldAccessor.cs
+++ b/Orm/Xtensive.Orm/Orm/Internals/FieldAccessors/EntityFieldAccessor.cs
@@ -1,4 +1,4 @@
-// Copyright (C) 2008-2020 Xtensive LLC.
+// Copyright (C) 2008-2021 Xtensive LLC.
 // This code is distributed under MIT license terms.
 // See the License.txt file in the project root for more information.
 // Created by: Alexey Gamzov