Following benchmarks show the improvement:

```
|             Method |     Mean |    Error |   StdDev |      Min |      Max |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------- |---------:|---------:|---------:|---------:|---------:|-------:|-------:|------:|----------:|
| SessionConstructor | 44.32 us | 0.251 us | 0.210 us | 43.86 us | 44.63 us | 0.3000 | 0.1000 |     - |     47 KB |
New Session constructor
|             Method |     Mean |    Error |   StdDev |      Min |      Max | Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------- |---------:|---------:|---------:|---------:|---------:|------:|------:|------:|----------:|
| SessionConstructor | 11.98 us | 0.127 us | 0.119 us | 11.79 us | 12.21 us |     - |     - |     - |     12 KB |
```